doc_retriever/app/requirements.txt
requests
langchain
qdrant-client
python-dotenv
langchain-openai
PyPDF2
pdfminer.six
pdfplumber
scikit-learn
pypdf